devtools/cdk/pyproject.toml
[project]
name = "adminer-bigquery-cdk"
version = "0.1.0"
description = "Adminer BigQuery CDK Stack"
requires-python = ">=3.8"
dependencies = [
    "aws-cdk-lib>=2.0.0",
    "constructs>=10.0.0",
    "boto3>=1.26.0",
]

[project.urls]
Homepage = "https://github.com/takemi-ohama/adminer"
Repository = "https://github.com/takemi-ohama/adminer.git"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["lib", "config"]

[dependency-groups]
dev = []